Helmet-and-Number-Plate-Detection-main/Major Project/requirements.txt
absl-py==0.13.0
anyio==3.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.4
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==24.2.0
autobahn==19.11.2
Automat==22.10.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==3.3.0
cachetools==4.2.2
certifi==2024.8.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.4
constantly==23.10.4
cryptography==42.0.8
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.4
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.2.2
ffmpeg==1.4
filelock==3.0.12
flatbuffers==1.12
fsspec==2024.6.1
gast==0.4.0
gdown==5.2.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.34.1
h11==0.14.0
h5py==3.1.0
hyperlink==21.0.0
idna==3.2
imageio==2.9.0
importlib-metadata==4.6.1
imutils==0.5.4
incremental==22.10.0
iniconfig==1.1.1
intel-openmp==2021.4.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
keras==2.7.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiteconnect==5.0.1
kiwisolver==1.3.1
libclang==12.0.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mkl==2021.4.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==6.1.0
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
notebook==6.4.0
numpy==1.26.4
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
opencv-python==4.10.0.84
openpyxl==3.0.7
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==24.1
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
patsy==0.5.1
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
plotly==4.14.3
pluggy==0.13.1
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protobuf==3.17.3
psutil==6.0.0
py==1.10.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycodestyle==2.7.0
pycparser==2.20
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.9.0
Pymem==1.8.5
PyMsgBox==1.0.9
pyOpenSSL==24.1.0
pyotp==2.9.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.27
PySocks==1.7.1
pytesseract==0.3.13
pytest==6.2.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-mnist==0.7
python-speech-features==0.6
PyTweening==1.0.3
pytz==2021.1
PyWavelets==1.1.1
pywin32==301
pywinpty==1.1.2
PyYAML==6.0.2
pyzmq==22.1.0
requests==2.26.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scenedetect==0.6.4
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.13.1
seaborn==0.11.1
selenium==4.24.0
Send2Trash==1.5.0
service-identity==24.1.0
six==1.15.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.1
statsmodels==0.12.2
sympy==1.13.1
tbb==2021.13.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.7.30
toml==0.10.2
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
tornado==6.1
tqdm==4.66.4
traitlets==5.0.5
trio==0.26.0
trio-websocket==0.11.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
ultralytics==8.3.2
ultralytics-thop==2.0.8
urllib3==1.26.6
virtualenv==20.4.7
wcwidth==0.2.5
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.0.1
wrapt==1.12.1
wsproto==1.2.0
xlrd==2.0.1
youtube-dl==2021.12.17
zipp==3.5.0
zope.interface==6.4.post2
